avoiding a roundtrip to AWS during configuration
`aws.accountId` is not property but a function call which may take up to couple of seconds in the configuration phase. this change will delay reading the `accountId` just before the deployment. as the account id is only fetched before the execution of the task then it's safe to remove the AWS credentials check